refactor(charts): type, util 파일 분리
