+++ apps/monitor-web/src/components/charts/index.ts
@@ -1,2 +1,1 @@
 export { default as ApiResponseTimeChart } from "./ApiResponseTimeChart";
-export type { ApiResponseTimeData } from "./types";
