+++ apps/monitor-web/src/types/index.ts
@@ -1,2 +1,7 @@
 export type { ApiStatus } from "./ApiStatusType";
+export type {
+  ApiResponseTimeData,
+  ChartDotProps,
+  ChartTooltipProps,
+} from "./ApiResponseTimeChartType";
 export type { ToastType, Toast } from "./ToastType";
